pyproject.toml
[project]
name = "demo-ai-stack"
version = "1.1.0"
description = "BMAD + CTO Guard governance toolkit"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6,<7",
]

[tool.ruff]
line-length = 100

target-version = "py310"
